Create separate folder for dynamic_type headers
