+++ csrc/polymorphic_value.h
@@ -19,7 +19,7 @@
 
 #define DYNAMIC_TYPE_CHECK TORCH_INTERNAL_ASSERT
 
-#include <dynamic_type.h>
+#include <dynamic_type/dynamic_type.h>
 #include <macros.h>
 #include <opaque_type.h>
