+++ lib/dynamic_type/test/ForAllTypes.cpp
@@ -6,7 +6,7 @@
  */
 // clang-format on
 
-#include "type_traits.h"
+#include "dynamic_type/type_traits.h"
 
 #include <tuple>
 #include <type_traits>
