+++ lib/dynamic_type/test/opcheck.cpp
@@ -6,7 +6,7 @@
  */
 // clang-format on
 
-#include "type_traits.h"
+#include "dynamic_type/type_traits.h"
 
 #include <memory>
