+++ lib/dynamic_type/test/print.cpp
@@ -9,7 +9,7 @@
 #include <gmock/gmock-matchers.h>
 #include <gtest/gtest.h>
 
-#include "dynamic_type.h"
+#include "dynamic_type/dynamic_type.h"
 
 #include <sstream>
 #include <string>
