+++ lib/dynamic_type/test/utils.h
@@ -9,7 +9,7 @@
 
 #include <gtest/gtest.h>
 
-#include "dynamic_type.h"
+#include "dynamic_type/dynamic_type.h"
 
 using namespace dynamic_type;
